+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.5.4](https://github.com/serlo/api.serlo.org/compare/v0.5.3..v0.5.4) - July 6, 2020
+
+### Fixed
+
+- Fix build
+
+## [0.5.3](https://github.com/serlo/api.serlo.org/compare/v0.5.2..v0.5.3) - July 6, 2020
+
+### Fixed
+
+- Fix navigation contract tests
+
 ## [0.5.2](https://github.com/serlo/api.serlo.org/compare/v0.5.1..v0.5.2) - June 5, 2020
 
 ### Fixed
@@ -16,7 +28,7 @@ All notable changes to this project will be documented in this file.
 
 ### Fixed
 
-- Ouput url-encoded aliases
+- Output url-encoded aliases
 
 ## [0.5.0](https://github.com/serlo/api.serlo.org/compare/v0.4.3..v0.5.0) - April 27, 2020